bot/requirements.txt
python-telegram-bot>=20.0
httpx>=2.28.0
ollama>=0.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pypdf2>=3.0.0